requirements.txt
fastapi
uvicorn
requests
pydantic-settings
python-dotenv
psycopg2-binary
redis
